+++ data-raw/eeda_test_data.R
@@ -64,8 +64,6 @@ events <-
   ) %>%
   dplyr::mutate(rfm = efe::quantile_of_quantiles(., c("recency", "frequency", "monetary")))
 
-
-
 # data set ----
 eeda_test_data <- tidyr::tibble(
   target = factor(sample(c("positive", "negative"), n_rows, replace = TRUE, prob = c(0.1, 0.9))),
@@ -88,6 +86,8 @@ eeda_test_data <- tidyr::tibble(
   eg_logical_na = eg_logical,
   eg_continuous = as.numeric(NA),
   eg_continuous_na = eg_continuous,
+  eg_cluster = sample(10*(1:7), n_rows, replace = TRUE) + runif(n_rows),
+  eg_cluster_na = eg_cluster,
   eg_integer_7 = sample(c(47, 53, 59, 61, 67, 71, 73), n_rows, replace = TRUE),
   eg_integer_7_na = eg_integer_7,
   eg_integer_50 = sample(1:50, n_rows, replace = TRUE),
@@ -116,6 +116,7 @@ eeda_test_data <- eeda_test_data %>%
 
     # add NA values
     eg_continuous_na = replace(eg_continuous, sample(1:n_rows, 200), NA),
+    eg_cluster_na = replace(eg_cluster, sample(1:n_rows, 200), NA),
     eg_factor_4_na = replace(eg_factor_4, sample(1:nrow(eeda_test_data), 200), NA),
     eg_factor_12_na = replace(eg_factor_12, sample(1:nrow(eeda_test_data), 200), NA),
     eg_factor_50_na = replace(eg_factor_50, sample(1:nrow(eeda_test_data), 200), NA),
@@ -134,3 +135,220 @@ eeda_test_data <- eeda_test_data %>%
 
 
 usethis::use_data(eeda_test_data, overwrite = TRUE)
+
+
+# x = (eeda_test_data[["eg_cluster"]])
+# suggest_number_of_clusters(eeda_test_data["eg_cluster"], diagnostic_file_prefix = "foo")
+# ?kmeans
+#
+# require(graphics)
+#
+# # a 2-dimensional example
+# x <- rbind(matrix(rnorm(100, sd = 0.3), ncol = 2),
+#            matrix(rnorm(100, mean = 1, sd = 0.3), ncol = 2))
+# colnames(x) <- c("x", "y")
+# (cl <- kmeans(x, 5))
+# plot(x, col = cl$cluster)
+# points(cl$centers, col = 1:2, pch = 8, cex = 2)
+#
+# # sum of squares
+# ss <- function(x) sum(scale(x, scale = FALSE)^2)
+#
+# ## cluster centers "fitted" to each obs.:
+# fitted.x <- fitted(cl);  head(fitted.x)
+# resid.x <- x - fitted(cl)
+#
+# ## Equalities : ----------------------------------
+# cbind(cl[c("betweenss", "tot.withinss", "totss")], # the same two columns
+#       c(ss(fitted.x), ss(resid.x),    ss(x)))
+# stopifnot(all.equal(cl$ totss,        ss(x)),
+#           all.equal(cl$ tot.withinss, ss(resid.x)),
+#           ## these three are the same:
+#           all.equal(cl$ betweenss,    ss(fitted.x)),
+#           all.equal(cl$ betweenss, cl$totss - cl$tot.withinss),
+#           ## and hence also
+#           all.equal(ss(x), ss(fitted.x) + ss(resid.x))
+# )
+#
+# kmeans(x,1)$withinss # trivial one-cluster, (its W.SS == ss(x))
+#
+# ## random starts do help here with too many clusters
+# ## (and are often recommended anyway!):
+# (cl <- kmeans(x, 5, nstart = 25))
+# plot(x, col = cl$cluster)
+# points(cl$centers, col = 1:5, pch = 8)
+# cut(x)
+#
+#
+#
+# hclust
+# x = dplyr::arrange((eeda_test_data["eg_cluster"]), eg_cluster)
+# foo(x)
+#
+# foo <- function(x,
+#                                        k_limit = 10,
+#                                        diagnostic_file_prefix = NULL) {
+#
+#   if (!is.null(diagnostic_file_prefix))
+#     stopifnot("A diagnostic file prefix was specified in suggest_number_of_clusters, but the /temp directory does not exist." = dir.exists("temp"))
+#
+#   if(sum(is.na(x)) > 0) {
+#     message(glue::glue("NA values in {names(x)} encountered and removed in suggest_number_of_clusters()"))
+#     x <- stats::na.omit(x)
+#   }
+#
+#   x <- unlist(x)
+#
+#   max_number_of_clusters <- pmin(length(unique(x)), k_limit)
+#
+#   # wss <- (nrow(x) - 1) * var(x)
+#   # # wss[1] is set above when wss is created
+#   # for (i in 2:max_number_of_clusters) {
+#   #   wss[i] <- sum(kmeans(x, centers = i)$withinss)
+#   # }
+#
+#   dist_x <- dist(x)
+#
+#   wss <- numeric(max_number_of_clusters)
+#   for (i in 1:max_number_of_clusters) {
+#     # wss[i] <- sum(stats::kmeans(x, centers = i)$withinss)
+#     # wss[i] <- stats::kmeans(x, centers = i)$tot.withinss
+#
+#     k_result <- stats::kmeans(x, centers = i, nstart = 25, iter.max = 15)
+#     wss[i] <- cluster::silhouette(k_result$cluster, dist_x)
+#   }
+#
+#   distance_from_diagonal <- numeric(max_number_of_clusters)
+#
+#   for (i in 1:max_number_of_clusters) {
+#     distance_from_diagonal[i] <- abs(scale_min_max(1:max_number_of_clusters)[i] - scale_min_max(wss[1:max_number_of_clusters])[i])
+#   }
+#
+#   k_best <- which.min(distance_from_diagonal)
+#
+#     graphics::plot(scale_min_max(1:max_number_of_clusters), scale_min_max(wss[1:max_number_of_clusters]), col = "gray60", cex = 0, type = "b", xlab = NA, ylab = NA)
+#     graphics::text(scale_min_max(1:max_number_of_clusters), scale_min_max(wss[1:max_number_of_clusters]), col = "royalblue4", labels = 1:max_number_of_clusters)
+#     graphics::abline(0, 1, col = "gray60")
+#
+#   return(k_best)
+# }
+#
+# factoextra::fviz_nbclust(x, kmeans, method = "wss")
+# factoextra::fviz_nbclust(x, kmeans, method = "silhouette")
+#
+# pam_limited <- function(x,k) list(cluster = cluster::pam(x,k, cluster.only=TRUE))
+#
+#
+# gap_stat <- cluster::clusGap(x, FUN = kmeans, K.max= max_number_of_clusters , B = 50)
+# q = factoextra::fviz_gap_stat(gap_stat)
+#
+#
+# plot(q)
+# q$
+# plot(hclust(dist(x)))
+#
+#
+#
+#
+# z = eeda_test_data[["eg_cluster"]]
+# y = eeda_test_data[["eg_cluster"]]
+#
+# plot(density(y))
+# plot(y)
+# plot(sort(y))
+#
+# x <- c(rnorm(50, sd=0.3), rnorm(50, mean=1, sd=0.3), rnorm(50, mean=2, sd=0.3))
+# k <- 8 # Divide x into 3 clusters
+#
+# result <- Ckmeans.1d.dp::Ckmeans.1d.dp(x, k)
+# result <- Ckmeans.1d.dp::Ckmeans.1d.dp(y, k)
+# colors <- RColorBrewer::brewer.pal(k, "Dark2")
+# plot(result, col.clusters = colors)
+#
+#
+#
+# plot(sort(y))
+# hist(y)
+# # Divide x into k clusters, k automatically selected (default: 1~9)
+# result <- Ckmeans.1d.dp::Ckmeans.1d.dp(y)
+# k <- max(result$cluster)
+# colors <- brewer.pal(k, "Dark2")
+# plot(result, col.clusters = colors)
+#
+# library(Ckmeans.1d.dp)
+#
+# # Ex. 1 The number of clusters is provided.
+# # Generate data from a Gaussian mixture model of three components
+# x = sort(y)
+# x <- c(rnorm(50, sd=0.2), rnorm(50, mean=1, sd=0.3), rnorm(100,
+#                                                            mean=-1, sd=0.25))
+# plot(x)
+# hist(x)
+# # Divide x into 3 clusters
+# k <- 3
+#
+# result <- Ckmedian.1d.dp(x)
+#
+# plot(result, main="Optimal univariate k-median given k")
+#
+# result <- Ckmeans.1d.dp(x, k)
+#
+# plot(result, main="Optimal univariate k-means given k")
+#
+# plot(x, col=result$cluster, pch=result$cluster, cex=1.5,
+#      main="Optimal univariate k-means clustering given k",
+#      sub=paste("Number of clusters given:", k))
+# abline(h=result$centers, col=1:k, lty="dashed", lwd=2)
+# legend("bottomleft", paste("Cluster", 1:k), col=1:k, pch=1:k,
+#        cex=1.5, bty="n")
+#
+# # Ex. 2 The number of clusters is determined by Bayesian
+# #       information criterion
+# # Generate data from a Gaussian mixture model of three components
+# x <- c(rnorm(50, mean=-3, sd=1), rnorm(50, mean=0, sd=.5),
+#        rnorm(50, mean=3, sd=1))
+# # Divide x into k clusters, k automatically selected (default: 1~9)
+#
+# result <- Ckmedian.1d.dp(x)
+#
+# plot(result, main="Optimal univariate k-median with k estimated")
+#
+# result <- Ckmeans.1d.dp(x)
+#
+# plot(result, main="Optimal univariate k-means with k estimated")
+#
+# k <- max(result$cluster)
+# plot(x, col=result$cluster, pch=result$cluster, cex=1.5,
+#      main="Optimal univariate k-means clustering with k estimated",
+#      sub=paste("Number of clusters is estimated to be", k))
+# abline(h=result$centers, col=1:k, lty="dashed", lwd=2)
+# legend("topleft", paste("Cluster", 1:k), col=1:k, pch=1:k,
+#        cex=1.5, bty="n")
+#
+# # Ex. 3 Segmenting a time course using optimal weighted
+# #       univariate clustering
+# n <- 160
+# t <- seq(0, 2*pi*2, length=n)
+# n1 <- 1:(n/2)
+# n2 <- (max(n1)+1):n
+# y1 <- abs(sin(1.5*t[n1]) + 0.1*rnorm(length(n1)))
+# y2 <- abs(sin(0.5*t[n2]) + 0.1*rnorm(length(n2)))
+# y <- c(y1, y2)
+#
+# w <- y^8 # stress the peaks
+# res <- Ckmeans.1d.dp(t, k=c(1:10), w)
+# plot(res)
+# plot(t, w, main = "Time course weighted k-means",
+#      col=res$cluster, pch=res$cluster,
+#      xlab="Time t", ylab="Transformed intensity w",
+#      type="h")
+# abline(v=res$centers, col="chocolate", lty="dashed")
+# text(res$centers, max(w) * .95, cex=0.5, font=2,
+#      paste(round(res$size / sum(res$size) * 100), "/ 100"))
+# [Package Ckmeans.1d.dp version 4.3.4 Index]
+#
+#
+#
+# BAMMtools::getJenksBreaks(y,9)
+#
+#