+++ NEWS.md
@@ -4,8 +4,10 @@
 
 # eeda 0.0.0.9002
 
-* Univariate functionality
+* Added key fields to eeda_test_data
+* Bivariate functionality added: plot_conditional_distribution and plot_conditional_distribution_by_category
 
 # eeda 0.0.0.9001
 
+* Univariate functionality
 * Code imported from disparate sources to be rationalized into a package.